refactor: rename `OnlineIssueTooManyDNSLookups`
